been debuging the rust intergration ot he godot frontend. Got it to passing the offers and answers around correctly(atleast for user ones) but it's not establishing the connection.  looked like I want' adding the ICE so this si my first atempt to fix that

+++ rust-logic/almeta-p2p/src/ice.rs
@@ -6,12 +6,12 @@ pub struct ICE {
     #[serde(rename = "Media")]
     pub media: String,
     #[serde(rename = "Index")]
-    pub index: i32, 
+    pub index: i64, 
     #[serde(rename = "Name")]
     pub name: String,
 }
 impl ICE {
-    pub fn new(media: String, index: i32, name: String) -> Self {
+    pub fn new(media: String, index: i64, name: String) -> Self {
         Self{media, index, name}
     }
 }
